clean up tests; create caf generation package; enable terra data tables pulling
