+++ requirements.txt
@@ -11,4 +11,7 @@ tqdm
 google-cloud-storage
 psutil
 setuptools
+
+# for cohort allele frequency generation
+firecloud
 pandas
